+++ localflavor/no/no_municipalities.py
@@ -1,25 +1,21 @@
 #: An alphabetical list of Norwegian municipalities (fylker) for
 #: use as `choices` in a formfield.
 MUNICIPALITY_CHOICES = (
+    ('agder', 'Agder'),
     ('akershus', 'Akershus'),
-    ('austagder', 'Aust-Agder'),
     ('buskerud', 'Buskerud'),
     ('finnmark', 'Finnmark'),
-    ('hedmark', 'Hedmark'),
-    ('hordaland', 'Hordaland'),
+    ('innlandet', 'Innlandet'),
     ('janmayen', 'Jan Mayen'),
     ('moreogromsdal', 'Møre og Romsdal'),
-    ('nordtrondelag', 'Nord-Trøndelag'),
     ('nordland', 'Nordland'),
-    ('oppland', 'Oppland'),
     ('oslo', 'Oslo'),
+    ('ostfold', 'Østfold'),
     ('rogaland', 'Rogaland'),
-    ('sognogfjordane', 'Sogn og Fjordane'),
     ('svalbard', 'Svalbard'),
-    ('sortrondelag', 'Sør-Trøndelag'),
     ('telemark', 'Telemark'),
     ('troms', 'Troms'),
-    ('vestagder', 'Vest-Agder'),
+    ('trondelag', 'Trøndelag'),
     ('vestfold', 'Vestfold'),
-    ('ostfold', 'Østfold')
+    ('vestland', 'Vestland')
 )