+++ docs/changelog.rst
@@ -17,6 +17,9 @@ Modifications to existing flavors:
 - Fix Belarus passport field description punctuation
   (`gh-484 <https://github.com/django/django-localflavor/pull/484>`_).
 - Change `Kiev` to `Kyiv` 🇺🇦 according to ISO_3166-2:UA
+- Update list of Norwegian counties(fylker) `according to changes at the start of 2024 <https://en.wikipedia.org/wiki/Counties_of_Norway#List_of_counties>`_
+  The changed counties are as follows: austagder and vestagder has been removed and combined into agder. hordaland and sognogfjordane has been removed and combined into vestland. nordtrondelag and sortrondelag has been removed and combined into trondelag. hedmark and oppland has been removed and combined into innlandet
+  (`gh-509 <https://github.com/django/django-localflavor/pull/509/files>`_).
 
 Other changes: